requirements.txt
networkx==2.6.3
numpy==1.21.6
osmnx==1.1.2
pandas==1.3.5
geopandas==0.10.2
geopy==2.4.1
Pillow==9.5.0
rtree==1.0.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.13.1
torch_geometric==2.3.1
torchvision==0.14.1
tqdm==4.66.1
